Feature/config utilities (#179)
* (wip) start moving to config utilities

* (wip) start moving config declarations to source files and breaking down cmake

* (wip) start on loop closure config conversion

* use factory for graph extractor creation

* fix room config

* punt on lcd config issues

* clean up unused header

* get tests to compile again

* start moving modules to factory based creation

* declare reconstruction and backend for factory

* add common module interface

* simplify backend constructor and fix minor issues

* add more necessary infrastructure for pipeline

* export config dep

* allow for graph extractor construction in combo integrator

* fix compilation issues after rebase

* move core pipeline over from ros

* drop explicit boost usage for custom factor

* split registration solution struct away from common

* change reconstruction callback signature

* (wip) start fixing module factories

* fix backend config issues

* fix config printing

* add logging for pipeline and configs

* fix graph extractor creation

* fix loop closure parsing

* reorganize config files and adapt to new parsing format

* copy reconstruction configs over for now

* add lcd config copy to default robot

* fix config file name

* fix object label namespacing

* check if graph extractor is valid before clearing

* add default values to graph extractor config

* drop unused file

+++ src/reconstruction/mesh_integrator.cpp
@@ -100,7 +100,7 @@ void MeshIntegrator::launchThreads(const BlockIndexList& blocks, bool interior_p
       new MixedThreadSafeIndex(blocks.size()));
 
   std::list<std::thread> threads;
-  for (size_t i = 0; i < config_.integrator_threads; ++i) {
+  for (int i = 0; i < config_.integrator_threads; ++i) {
     if (interior_pass) {
       threads.emplace_back(
           &MeshIntegrator::processInterior, this, blocks, index_getter.get());
